# - Criando toda rotina para salvar trade e consultar trades cadastrados

+++ src/UBSAg.Negotiations.Domain/Constants/RiskValueConstant.cs
@@ -2,6 +2,6 @@
 {
     public static class RiskValueConstant
     {
-        public const double RISK_CUTOFF_VALUE = 1_000_000;
+        public const decimal RISK_CUTOFF_VALUE = 1_000_000;
     }
 }
