+++ src/UBSAg.Negotiations.Domain/Entities/Trade.cs
@@ -4,10 +4,10 @@ namespace UBSAg.Negotiations.Domain.Entities
 {
     public class Trade : ITrade
     {
-        public double Value { get; }
+        public decimal Value { get; }
         public string ClientSector { get; }        
 
-        public Trade(double value, string clientSector)
+        public Trade(decimal value, string clientSector)
         {
             Value=value;
             ClientSector=clientSector;
